@@ -6,6 +6,6 @@
 ///   https://github.com/BarthaBRW/shared_preferences_settings
 library flutter_settings_screens;
 
-export 'package:flutter_settings_screens/src/cache/cache.dart';
-export 'package:flutter_settings_screens/src/settings.dart';
-export 'package:flutter_settings_screens/src/settings_screen.dart';
+export 'src/cache/cache.dart';
+export 'src/settings.dart';
+export 'src/widgets/settings_widgets.dart';
